pyproject.toml
[project]
name = "sleep-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "supabase>=2.16.0",
]
